Extract sdl_backend in its own module.
